+++ WebHook.SubscriptionStore.Client/ISubscriptionStore.cs
@@ -5,7 +5,7 @@ namespace WebHook.SubscriptionStore.Client;
 
 public interface ISubscriptionStore
 {
-    // TODO: async
+    // TODO: async?
     IReadOnlyList<Subscription> GetSubscriptionsFor<TEvent>(TEvent @event, CancellationToken cancellationToken) where TEvent : IEvent;
     bool IsActive(int subscriptionId);
 }
